Reactive Microservices

+++ microservices/recommendation-service/src/main/java/nl/paulzijlmans/microservices/core/recommendation/services/RecommendationServiceImpl.java
@@ -1,7 +1,5 @@
 package nl.paulzijlmans.microservices.core.recommendation.services;
 
-import java.util.List;
-
 import nl.paulzijlmans.api.core.recommendation.Recommendation;
 import nl.paulzijlmans.api.core.recommendation.RecommendationService;
 import nl.paulzijlmans.api.exceptions.InvalidInputException;
@@ -13,6 +11,10 @@ import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.dao.DuplicateKeyException;
 import org.springframework.web.bind.annotation.RestController;
+import reactor.core.publisher.Flux;
+import reactor.core.publisher.Mono;
+
+import static java.util.logging.Level.FINE;
 
 @RestController
 public class RecommendationServiceImpl implements RecommendationService {
@@ -20,9 +22,7 @@ public class RecommendationServiceImpl implements RecommendationService {
     private static final Logger LOG = LoggerFactory.getLogger(RecommendationServiceImpl.class);
 
     private final RecommendationRepository repository;
-
     private final RecommendationMapper mapper;
-
     private final ServiceUtil serviceUtil;
 
     @Autowired
@@ -33,37 +33,44 @@ public class RecommendationServiceImpl implements RecommendationService {
     }
 
     @Override
-    public Recommendation createRecommendation(Recommendation body) {
-        try {
-            RecommendationEntity entity = mapper.apiToEntity(body);
-            RecommendationEntity newEntity = repository.save(entity);
+    public Mono<Recommendation> createRecommendation(Recommendation body) {
 
-            LOG.debug("createRecommendation: created a recommendation entity: {}/{}", body.productId(), body.recommendationId());
-            return mapper.entityToApi(newEntity);
-
-        } catch (DuplicateKeyException dke) {
-            throw new InvalidInputException("Duplicate key, Product Id: " + body.productId() + ", Recommendation Id:" + body.recommendationId());
+        if (body.productId() < 1) {
+            throw new InvalidInputException("Invalid productId: " + body.productId());
         }
+
+        RecommendationEntity entity = mapper.apiToEntity(body);
+
+        return repository.save(entity)
+                .log(LOG.getName(), FINE)
+                .onErrorMap(
+                        DuplicateKeyException.class,
+                        ex -> new InvalidInputException("Duplicate key, Product Id: " + body.productId() + ", Recommendation Id:" + body.recommendationId()))
+                .map(mapper::entityToApi);
     }
 
     @Override
-    public List<Recommendation> getRecommendations(int productId) {
+    public Flux<Recommendation> getRecommendations(int productId) {
 
         if (productId < 1) {
             throw new InvalidInputException("Invalid productId: " + productId);
         }
 
-        List<RecommendationEntity> entityList = repository.findByProductId(productId);
-        List<Recommendation> list = entityList.stream().map(e -> new Recommendation(productId, e.getRecommendationId(), e.getAuthor(), e.getRating(), e.getContent(), serviceUtil.getServiceAddress())).toList();
+        LOG.info("Will get recommendations for product with id={}", productId);
 
-        LOG.debug("getRecommendations: response size: {}", list.size());
-
-        return list;
+        return repository.findByProductId(productId)
+                .log(LOG.getName(), FINE)
+                .map(e -> new Recommendation(productId, e.getRecommendationId(), e.getAuthor(), e.getRating(), e.getContent(), serviceUtil.getServiceAddress()));
     }
 
     @Override
-    public void deleteRecommendations(int productId) {
+    public Mono<Void> deleteRecommendations(int productId) {
+
+        if (productId < 1) {
+            throw new InvalidInputException("Invalid productId: " + productId);
+        }
+
         LOG.debug("deleteRecommendations: tries to delete recommendations for the product with productId: {}", productId);
-        repository.deleteAll(repository.findByProductId(productId));
+        return repository.deleteAll(repository.findByProductId(productId));
     }
 }
